+++ sdk/securitycenter/Azure.ResourceManager.SecurityCenter/CHANGELOG.md
@@ -1,12 +1,14 @@
 # Release History
 
-## 1.0.0-beta.1 (Unreleased)
+## 1.2.0-beta.8 (Unreleased)
 
 ### Features Added
 
-- Initial release of the Azure Resource Manager Security client library generated from TypeSpec.
+- Regenerated the Azure Resource Manager Security client library from TypeSpec.
 
 ### Breaking Changes
+
+- Removed APIs for service operations that are no longer described by the current SecurityCenter TypeSpec.
 
 ### Bugs Fixed